Consolidate various submitCatalogLoginForm methods to DemoDriverTestTrait.
